refactor(java): Detach MemoryBuffer from Fury
